refactor: share makeTagValueDefinition between fixture and mocks; job/city tag fixtures
